# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for ليلتك wedding venue platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List venues with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: minCapacity
          in: query
          schema:
            type: integer
        - name: city
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueListResponse"
...
  /venues/nearby:
    get:
      operationId: getNearbyVenues
      tags: [venues]
      summary: Get venues near Riyadh districts
      parameters:
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Nearby venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
...
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get single venue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueDetail"
        "404":
          description: Not found
  /venues/{id}/availability:
    get:
      operationId: getVenueAvailability
      tags: [venues]
      summary: Get available dates for a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueAvailability"

  # ─── BOOKINGS ────────────────────────────────────────────────────────────────
...
  /venues/{id}/reviews:
    get:
      operationId: listVenueReviews
      tags: [reviews]
      summary: Get reviews for a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createVenueReview
      tags: [reviews]
      summary: Post a review for a venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── MERCHANTS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Venue:
      type: object
      required: [id, nameAr, categorySlug, pricePerNight, capacity, district, rating, reviewCount, images, isFeatured, isAvailable]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        categorySlug:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        district:
          type: string
          description: Riyadh district in Arabic
        rating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        merchantId:
          type: ["integer", "null"]
    VenueDetail:
      type: object
      required: [id, nameAr, categorySlug, pricePerNight, capacity, district, rating, reviewCount, images, isFeatured, isAvailable, amenities]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        categorySlug:
          type: string
        pricePerNight:
          type: number
        capacity:
          type: integer
        district:
          type: string
        rating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        isAvailable:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        services:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        merchantId:
          type: ["integer", "null"]
        menuPdf:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        socialLinks:
          type: ["object", "null"]
          properties:
            instagram:
              type: ["string", "null"]
            twitter:
              type: ["string", "null"]
    VenueListResponse:
      type: object
      required: [venues, total, page, limit]
      properties:
        venues:
          type: array
          items:
            $ref: "#/components/schemas/Venue"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    VenueAvailability:
      type: object
      required: [venueId, availableDates, bookedDates]
      properties:
        venueId:
          type: integer
        availableDates:
          type: array
          items:
            type: string
        bookedDates:
          type: array
          items:
            type: string
...
    Review:
      type: object
      required: [id, venueId, authorName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        venueId:
          type: integer
        authorName:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [reviewToken, authorName, rating, comment]
      properties:
        reviewToken:
          type: string
          description: Unguessable UUID token from the booking (proves ownership without account login)
        authorName:
          type: string
          minLength: 1
          maxLength: 100
        rating:
          type: number
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 1
          maxLength: 2000
...
  securitySchemes:
    ManageBearerAuth:
      type: http
      scheme: bearer
      description: >
        Invitation management bearer token (manageToken).
        Send as `Authorization: Bearer <token>`. Never place this token in
        a URL path or query string — it must only travel as a request header.