# lib/api-zod/src/generated/types/venueAvailability.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ليلتك wedding venue platform
 * OpenAPI spec version: 0.1.0
 */

export interface VenueAvailability {
  venueId: number;
  availableDates: string[];
  bookedDates: string[];
}