# lib/api-zod/src/generated/types/venueListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for ليلتك wedding venue platform
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from './venue';

export interface VenueListResponse {
  venues: Venue[];
  total: number;
  page: number;
  limit: number;
}